requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库
sqlalchemy==2.0.23

# OCR和图像处理
easyocr==1.7.0
pillow==9.5.0
opencv-python==4.8.1.78

# 翻译API
googletrans==4.0.0rc1
requests==2.31.0

# 数据处理
pandas==2.1.3
numpy==1.25.2

# 前端界面
streamlit==1.28.1
plotly==5.17.0

# 工具库
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1

# 开发工具
pytest==7.4.3
black==23.11.0 